requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
distro==1.9.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
milvus-lite==2.4.10
numpy==2.2.0
openai==1.58.1
pandas==2.2.3
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
pymilvus==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
